Refactor ConfigTable
