+++ app/src/main/java/nil/nadph/qnotified/hook/ReplyNoAtHook.java
@@ -22,7 +22,7 @@ import android.os.Looper;
 import android.widget.Toast;
 
 import de.robv.android.xposed.XC_MethodHook;
-import me.ketal.util.TIMConfigTable;
+import me.singleneuron.qn_kernel.tlb.TIMConfigTable;
 import me.singleneuron.qn_kernel.tlb.ConfigTable;
 import nil.nadph.qnotified.SyncUtils;
 import nil.nadph.qnotified.config.ConfigManager;
@@ -62,11 +62,7 @@ public class ReplyNoAtHook extends BaseDelayableHook {
     public boolean init() {
         if (inited) return true;
         try {
-            String method;
-            if (isTim(getApplication()))
-                method = TIMConfigTable.INSTANCE.getConfig(ReplyNoAtHook.class.getSimpleName());
-            else
-                method = ConfigTable.INSTANCE.getConfig(ReplyNoAtHook.class.getSimpleName());
+            String method = ConfigTable.INSTANCE.getConfig(ReplyNoAtHook.class.getSimpleName());
             /*int ver = getHostVersionCode32();
             if (ver >= 1630) {
                 method = "l";
